Update README.md
"A hacker has attacked the IT system of your school and destroyed the database. As a leader of the Science Club, your school has asked for your help. The situation is grave. There is only one week before the final annual exams. You need to recover the annual scores of the students. The school management provides the data from the archive, and the Science Club members help you with csv files. Your task is to create the school database and calculate the annual score."